refactor(core): decompose Python storage CLI
